исправлена ошибка маршрутов

+++ bot/handlers/topic.py
@@ -11,16 +11,18 @@ router = Router()
 @router.message(F.chat.id == settings.bot.GROUP_ID)
 async def handle_message_in_topic(message: Message):
     user = message.from_user
-    if message.message_thread_id and not user.is_bot:
+    topic_id = message.message_thread_id
+    tg_id = await UserTopics.get_topic_id_by_tg_id(topic_id)
+    if topic_id and not user.is_bot:
         try:
             # Текст
             if message.text:
-                await bot.send_message(chat_id=user.id, text=message.text)
+                await bot.send_message(chat_id=tg_id, text=message.text)
 
             # Фото
             elif message.photo:
                 await bot.send_photo(
-                    chat_id=user.id,
+                    chat_id=tg_id,
                     photo=message.photo[-1].file_id,  # Самое большое по размеру
                     caption=message.caption or ""
                 )
@@ -28,7 +30,7 @@ async def handle_message_in_topic(message: Message):
             # Документ
             elif message.document:
                 await bot.send_document(
-                    chat_id=user.id,
+                    chat_id=tg_id,
                     document=message.document.file_id,
                     caption=message.caption or ""
                 )
@@ -36,7 +38,7 @@ async def handle_message_in_topic(message: Message):
             # Видео
             elif message.video:
                 await bot.send_video(
-                    chat_id=user.id,
+                    chat_id=tg_id,
                     video=message.video.file_id,
                     caption=message.caption or ""
                 )
@@ -44,7 +46,7 @@ async def handle_message_in_topic(message: Message):
             # Голосовое сообщение
             elif message.voice:
                 await bot.send_voice(
-                    chat_id=user.id,
+                    chat_id=tg_id,
                     voice=message.voice.file_id,
                     caption=message.caption or ""
                 )
@@ -56,9 +58,13 @@ async def handle_message_in_topic(message: Message):
             print(f"\n❌ Ошибка при отправке сообщения пользователю: {e}\n")
 
 
-# Обработка сообщений от пользователя в ЛС
+# Обработка сообщений от пользователя в БОТЕ
 @router.message()
 async def handle_message_from_user(message: Message):
+
+    if message.text == '/broadcast':
+        return
+
     user_topic = await UserTopics.get_by_tg_id(message.from_user.id)
     if not user_topic:
         await message.delete()
